requirements.txt
Flask==2.0.1
line-bot-sdk==1.20.0
english-words==1.0.4